+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/VillageBaseStructure.java
@@ -62,7 +62,19 @@ public class VillageBaseStructure extends Structure<NoFeatureConfig> {
         public void init(DynamicRegistries dynamicRegistryManager, ChunkGenerator chunkGenerator, TemplateManager structureManager, int chunkX, int chunkZ, Biome biome, NoFeatureConfig NoFeatureConfig) {
 
             BlockPos blockpos = new BlockPos(chunkX * 16, 0, chunkZ * 16);
-            JigsawManager.method_30419(dynamicRegistryManager, new VillageConfig(() -> dynamicRegistryManager.get(Registry.TEMPLATE_POOL_WORLDGEN).getOrDefault(START_POOL), STRUCTURE_SIZE), AbstractVillagePiece::new, chunkGenerator, structureManager, blockpos, this.components, this.rand, true, true);
+            JigsawManager.method_30419(
+                    dynamicRegistryManager,
+                    new VillageConfig(() -> dynamicRegistryManager.get(Registry.TEMPLATE_POOL_WORLDGEN)
+                            .getOrDefault(START_POOL),
+                            STRUCTURE_SIZE),
+                    AbstractVillagePiece::new,
+                    chunkGenerator,
+                    structureManager,
+                    blockpos,
+                    this.components,
+                    this.rand,
+                    true,
+                    true);
             this.recalculateStructureSize();
         }
     }